sadf

+++ maze/maze/FileHandler.h
@@ -13,10 +13,10 @@ private:
 	bool					checkErrors();
 	inline void				pushError(ErrorType type, const string & str) { m_errors.list.push_back(Pair(type, str)); }
 	string					getName(string & line);
-	size_t					getIntValue(const string & input, const ErrorType error, string & line);
-	MazeBoard				getBoard(const size_t rows, const size_t cols, Coordinate & playerLocation, Coordinate & endLocation, string & line);
-	void					handleSpecialChar(const char c, Coordinate & location, const size_t i, const size_t j, bool & seenChar, string & line, const ErrorType e);
-	void					handleInvalidChar(const char c, const size_t i, const size_t j);
+	int					getIntValue(const string & input, const ErrorType error, string & line);
+	MazeBoard				getBoard(const int rows, const int cols, Coordinate & playerLocation, Coordinate & endLocation, string & line);
+	void					handleSpecialChar(const char c, Coordinate & location, const int i, const int j, bool & seenChar, string & line, const ErrorType e);
+	void					handleInvalidChar(const char c, const int i, const int j);
 
 public:
 	FileHandler(int argc, char* argv[]);
